% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_coxph.R
\name{estimate_coxph_tl}
\alias{estimate_coxph_tl}
\title{T-learner of Cox PH}
\usage{
estimate_coxph_tl(data, data.test, times)
}
\arguments{
\item{times}{The prediction time of interest}

\item{x}{The baseline covariates}

\item{w}{The treatment variable (0 or 1)}

\item{y}{The follow-up time}

\item{D}{The event indicator}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
T-learner, implemented via Cox proportional hazard models
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

tcoxph_cate = estimate_coxph_tl(data, data.test, times)
}
}
