% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf.R
\name{estimate_csf_probs}
\alias{estimate_csf_probs}
\title{causal survival forest}
\usage{
estimate_csf_probs(
  data,
  data.test,
  times,
  alpha = 0.05,
  ps = NULL,
  cen_fit = "KM"
)
}
\arguments{
\item{data}{The training data set}

\item{data.test}{The testing data set}

\item{times}{The prediction time of interest}

\item{alpha}{Imbalance tuning parameter for a split (see grf documentation)}

\item{ps}{The propensity score}

\item{cen_fit}{The choice of model fitting for censoring}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
causal survival forest, implemented via the grf R package
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

csf_cate = estimate_csf_probs(data, data.test, times, ps = 0.5)
}
}
