% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_grf.R
\name{estimate_grf_tl}
\alias{estimate_grf_tl}
\title{T-learner of grf}
\usage{
estimate_grf_tl(data, data.test, times, alpha = 0.05)
}
\arguments{
\item{data}{The training data set}

\item{data.test}{The testing data set}

\item{times}{The prediction time of interest}

\item{alpha}{Imbalance tuning parameter for a split (see grf documentation)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
T-learner, implemented via survival_forest (grf package)
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

tgrf_surv_cate = estimate_grf_tl(data, data.test, times)
}
}
