% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fgrf.R
\name{Fgrf}
\alias{Fgrf}
\title{F-learner of grf}
\usage{
Fgrf(x, w, y, weight, pscore = rep(0.5, nrow(x)), num.trees = 2000)
}
\arguments{
\item{x}{matrix of covariates}

\item{w}{vector of treatment indicators (0 or 1)}

\item{y}{vector of response values}

\item{weight}{vector of subject level weights}

\item{pscore}{vector of propensity scores}

\item{num.trees}{number of trees}
}
\value{
An Flasso object
}
\description{
F-learner, implemented via the grf package
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
weight <- rep(1, length(Y))
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

Fgrf_fit = Fgrf(data$X, data$W, data$Y, weight)
Fgrf_cate = predict(Fgrf_fit, data.test$X)
}
}
