% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgps.R
\name{generate_tutorial_survival_data}
\alias{generate_tutorial_survival_data}
\title{The following DGPs are available for benchmarking purposes, T is the failure time
and C the censoring time:}
\usage{
generate_tutorial_survival_data(
  n,
  p,
  p_b = NULL,
  p_i = NULL,
  f_b = NULL,
  f_i = NULL,
  times = NULL,
  pi = 0.5,
  beta = 1,
  gamma = 1,
  rho = 2,
  cen_scale = 4,
  cenM = "indX",
  Y.max = NULL,
  n.mc = 10000,
  dgp = "fcomplex"
)
}
\arguments{
\item{n}{The number of samples.}

\item{p}{The total number of covariates.}

\item{p_b}{The number of covariates in the main effect function.}

\item{p_i}{The number of covariates in the interaction term.}

\item{f_b}{Function form of the main effects (linear/nonlinear).}

\item{f_i}{Function form of the interaction term (linear/nonlinear).}

\item{times}{The time of interest.}

\item{pi}{The probability of receiving treatment.}

\item{beta}{The coefficient of main effect.}

\item{gamma}{The coefficient of treatment-covariate interaction.}

\item{rho}{The shape parameter of a Weibull distribution for censoring time.}

\item{cen_scale}{The scale parameter of a Weibull distribution for censoring time.}

\item{cenM}{The complexity of censoring mechanism (independent/dependent of covariates).}

\item{Y.max}{The maximum follow-up time (optional).}

\item{n.mc}{The number of monte carlo draws to estimate the treatment effect with. Default is 10000.}

\item{dgp}{The type of DGP.}
}
\value{
A list with entries:
\code{X}: the covariates, \code{Y}: the event times, \code{W}: the treatment indicator, \code{D}: the censoring indicator,
\code{catesp}: the treatment effect estimated by monte carlo, \code{catesp.sign}: the true sign of the cate for ITR comparison,
\code{dgp}: the dgp name, \code{Y.max}: the maximum follow-up time.
}
\description{
The following DGPs are available for benchmarking purposes, T is the failure time
and C the censoring time:
}
\examples{
\donttest{
# Generate data
n <- 1000
p <- 25
data <- generate_tutorial_survival_data(n, p, p_b = 1, p_i = 1, f_b = "L", f_i = "L", times = 0.2)
# Get true CATE in terms of difference in survival probabilities at time 0.2
cate <- generate_tutorial_survival_data(n, p, p_b = 1, p_i = 1, f_b = "L", f_i = "L", times = 0.2)$cate.sp
}

}
