% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fgrf.R
\name{predict.Fgrf}
\alias{predict.Fgrf}
\title{predict for Fgrf}
\usage{
\method{predict}{Fgrf}(object, newx, ...)
}
\arguments{
\item{object}{An Fgrf object}

\item{newx}{Covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{Additional arguments (currently not used)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
get estimated tau(x) using the trained Fgrf model
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
weight <- rep(1, length(Y))
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

Fgrf_fit = Fgrf(X, W, Y, weight)
Fgrf_cate = predict(Fgrf_fit, data.test$X)
}

}
