% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrf.R
\name{predict.rgrf}
\alias{predict.rgrf}
\title{predict for rgrf}
\usage{
\method{predict}{rgrf}(object, newx = NULL, tau_only = TRUE, meta_learner = TRUE, ...)
}
\arguments{
\item{object}{a rgrf object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{tau_only}{if set to TRUE, onlly return prediction on tau. Otherwise, return a list including prediction on tau, propensity score, and baseline main effect.}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained rgrf model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

rboost_fit = rboost(x, w, y)
rboost_est = predict(rboost_fit, x)
}


}
