% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{predict.rlasso}
\alias{predict.rlasso}
\title{predict for rlasso}
\usage{
\method{predict}{rlasso}(object, newx = NULL, ...)
}
\arguments{
\item{object}{An rlasso object}

\item{newx}{Covariate matrix to make predictions on. If null, return the tau(X) predictions on the training data}

\item{...}{Additional arguments (currently not used)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
get estimated tau(X) using the trained rlasso model
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

rlasso_fit = rlasso(X, W, Y, D, times, p_hat = 0.5)
rlasso_cate = predict(rlasso_fit, X)
}

}
