% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoxph.R
\name{predict.scoxph}
\alias{predict.scoxph}
\title{predict for scoxph}
\usage{
\method{predict}{scoxph}(object, newx = NULL, times, ...)
}
\arguments{
\item{object}{A scoxph object}

\item{newx}{Covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{times}{The prediction time of interest}

\item{...}{Additional arguments (currently not used)}
}
\value{
vector of estimated conditional average treatment effects
}
\description{
get estimated tau(x) using the trained scoxph model
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

scoxph_fit = scoxph(X, W, Y, D, times)
scoxph_cate = predict(scoxph_fit, X, times)
}
}
