% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slasso_surv.R
\name{predict.slasso_surv}
\alias{predict.slasso_surv}
\title{predict for slasso_surv}
\usage{
\method{predict}{slasso_surv}(object, newx = NULL, times, ...)
}
\arguments{
\item{object}{A slasso_surv object}

\item{newx}{Covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{times}{The prediction time of interest}

\item{...}{Additional arguments (currently not used)}
}
\value{
vector of estimated conditional average treatment effects
}
\description{
get estimated tau(x) using the trained slasso_surv model
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

slasso_surv_fit = slasso_surv(x, w, y, D, times)
slasso_surv_cate = predict(slasso_surv_fit, x, times)
}


}
