% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_fl_lasso.R
\name{predict.surv_fl_lasso}
\alias{predict.surv_fl_lasso}
\title{predict for surv_fl_lasso}
\usage{
\method{predict}{surv_fl_lasso}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An surv_fl_lasso object}

\item{newdata}{Covariate matrix to make predictions on. If null, return the tau(X) predictions on the training data}

\item{...}{Additional arguments (currently not used)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
get estimated tau(X) using the trained surv_fl_lasso model
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.fl.lasso.fit = surv_fl_lasso(X, W, Y, D, times, ps = 0.5)
cate = predict(surv.fl.lasso.fit)
cate.test = predict(surv.fl.lasso.fit, X.test)
}

}
