% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rl_lasso.R
\name{predict.surv_rl_lasso}
\alias{predict.surv_rl_lasso}
\title{Predict with a R-learner with Lasso}
\usage{
\method{predict}{surv_rl_lasso}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An surv_rl_lasso object}

\item{newdata}{Covariate matrix to make predictions on. If null, return the tau(X) predictions on the training data}

\item{...}{Additional arguments (currently not used)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
Obtain estimated tau(X) using a trained R-learner with Lasso model
}
\details{
Remark: CATE predictions can only be made at the time point used to define the outcome in the trained model
}
\examples{
\donttest{
n <- 1000; p <- 25
t0 <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[ ,1, drop = FALSE] + (-0.5 - 1 * X[ ,2, drop = FALSE]) * W)) ^ 2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC / (4 ^ 2)) ^ (1 / 2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.rl.lasso.fit <- surv_rl_lasso(X, Y, W, D, t0, W.hat = 0.5)
cate <- predict(surv.rl.lasso.fit)
cate.test <- predict(surv.rl.lasso.fit, X.test)
}

}
