% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_sl_lasso.R
\name{predict.surv_sl_lasso}
\alias{predict.surv_sl_lasso}
\title{predict for surv_sl_lasso}
\usage{
\method{predict}{surv_sl_lasso}(object, newdata = NULL, t0 = NULL, ...)
}
\arguments{
\item{object}{A surv_sl_lasso object}

\item{newdata}{Covariate matrix to make predictions on. If null, return the tau(X) predictions on the training data}

\item{t0}{The prediction time of interest}

\item{...}{Additional arguments (currently not used)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
get estimated tau(X) using the trained surv_sl_lasso model
}
\examples{
\donttest{
n <- 1000; p <- 25
t0 <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[ ,1] + (-0.5 - 1 * X[ ,2]) * W)) ^ 2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC / (4 ^ 2)) ^ (1 / 2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.sl.lasso.fit <- surv_sl_lasso(X, Y, W, D, t0)
cate <- predict(surv.sl.lasso.fit)
cate.test <- predict(surv.sl.lasso.fit, X.test)
}

}
