% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoxph.R
\name{scoxph}
\alias{scoxph}
\title{S-learner, implemented via glmnet (lasso)}
\usage{
scoxph(x, w, y, D, times)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{D}{event indicator}

\item{times}{prediction time of interest}
}
\description{
S-learner, as proposed by Imai and Ratkovic (2013), implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

slasso_fit = slasso(x, w, y)
slasso_est = predict(slasso_fit, x)
}
}
