% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_fl_grf.R
\name{surv_fl_grf}
\alias{surv_fl_grf}
\title{F-learner of grf}
\usage{
surv_fl_grf(X, W, Y, D, times, alpha = 0.05, ps = NULL, cen_fit = "KM")
}
\arguments{
\item{times}{The prediction time of interest}

\item{alpha}{Imbalance tuning parameter for a split (see grf documentation)}

\item{ps}{The propensity score}

\item{cen_fit}{The choice of model fitting for censoring}

\item{data}{The training data set}

\item{data.test}{The testing data set}

\item{newX}{The test data set (covariates only)}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
F-learner, implemented via survival_forest in the grf package
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
X.test <- matrix(rnorm(n * p), n, p)

surv_fl_grf_fit = surv_fl_grf(X, W, Y, D, times, ps = 0.5)
cate = predict(surv_fl_grf_fit)
cate.test = predict(surv_fl_grf_fit, X.test)
}
}
