% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rl_grf.R
\name{surv_rl_grf}
\alias{surv_rl_grf}
\title{R-learner of grf}
\usage{
surv_rl_grf(
  X,
  W,
  Y,
  D,
  times,
  ps = NULL,
  alpha = 0.05,
  cen_fit = "KM",
  newX = NULL
)
}
\arguments{
\item{times}{The prediction time of interest}

\item{ps}{The propensity score}

\item{alpha}{Imbalance tuning parameter for a split (see grf documentation)}

\item{cen_fit}{The choice of model fitting for censoring}

\item{newX}{The test data set (covariates only)}

\item{data}{The training data set}

\item{data.test}{The testing data set}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
R-learner, implemented via the grf package
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

cate = surv_rl_grf(X, W, Y, D, times, ps = 0.5, newX = X)
}
}
