% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rl_grf_lasso.R
\name{surv_rl_grf_lasso}
\alias{surv_rl_grf_lasso}
\title{R-learner of grf and lasso}
\usage{
surv_rl_grf_lasso(
  X,
  W,
  Y,
  D,
  times = NULL,
  k_folds = 10,
  p_hat = NULL,
  m_hat = NULL,
  c_hat = NULL,
  failure.times = NULL,
  num.trees = 2000,
  sample.weights = NULL,
  clusters = NULL,
  equalize.cluster.weights = FALSE,
  sample.fraction = 0.5,
  mtry = min(ceiling(sqrt(ncol(X)) + 20), ncol(X)),
  min.node.size = 5,
  honesty = TRUE,
  honesty.fraction = 0.5,
  honesty.prune.leaves = TRUE,
  alpha = 0.05,
  imbalance.penalty = 0,
  stabilize.splits = TRUE,
  ci.group.size = 2,
  tune.parameters = "none",
  compute.oob.predictions = TRUE,
  num.threads = NULL,
  seed = runif(1, 0, .Machine$integer.max),
  lambda_tau = NULL,
  lambda_choice = "lambda.min",
  penalty_factor = NULL,
  cen_fit = "KM",
  verbose = FALSE
)
}
\arguments{
\item{X}{The baseline covariates}

\item{W}{The treatment variable (0 or 1)}

\item{Y}{The follow-up time}

\item{D}{The event indicator}

\item{times}{The prediction time of interest}

\item{k_folds}{Number of folds for cross validation}

\item{p_hat}{Propensity score}

\item{m_hat}{Conditional mean outcome E(Y|X)}

\item{c_hat}{Censoring weights}

\item{failure.times}{A vector of event times to fit the survival curve at.}

\item{num.trees}{Number of trees grown in the forest}

\item{sample.weights}{See grf documentation}

\item{clusters}{See grf documentation}

\item{equalize.cluster.weights}{See grf documentation}

\item{sample.fraction}{See grf documentation}

\item{mtry}{See grf documentation}

\item{min.node.size}{See grf documentation}

\item{honesty}{See grf documentation}

\item{honesty.fraction}{See grf documentation}

\item{honesty.prune.leaves}{See grf documentation}

\item{alpha}{See grf documentation}

\item{imbalance.penalty}{See grf documentation}

\item{stabilize.splits}{See grf documentation}

\item{ci.group.size}{See grf documentation}

\item{tune.parameters}{See grf documentation}

\item{compute.oob.predictions}{See grf documentation}

\item{num.threads}{See grf documentation}

\item{seed}{See grf documentation}

\item{lambda_tau}{User-supplied lambda sequence for cross validation in the cate model}

\item{lambda_choice}{How to cross-validate; choose from "lambda.min" or "lambda.1se"}

\item{penalty_factor}{User-supplied penalty factor, must be of length the same as number of features in X}

\item{cen_fit}{The choice of model fitting for censoring}
}
\value{
a surv_rl_grf_lasso object
}
\description{
R-learner, implemented via the grf package for nuisance parameter estimation and lasso for target parameter
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
X.test <- matrix(rnorm(n * p), n, p)

surv_rl_grf_lasso_fit = surv_rl_grf_lasso(X, W, Y, D, times, p_hat = 0.5)
cate = predict(surv_rl_grf_lasso_fit)
cate.test = predict(surv_rl_grf_lasso_fit, X.test)
}
}
