% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_s_grf.R
\name{surv_s_grf}
\alias{surv_s_grf}
\title{S-learner for survival data implemented via grf::survival_forest}
\usage{
surv_s_grf(data, data.test, time, alpha = 0.05, propensity = 0.5)
}
\arguments{
\item{data}{training data}

\item{data.test}{testing data}

\item{time}{time at which to predict}
}
\description{
S-learner via grf::survival_forest
}
