% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survlearners-package.R
\docType{package}
\name{survlearners-package}
\alias{survlearners}
\alias{survlearners-package}
\title{survlearners: Survival Metalearners}
\description{
A package for...

etc

Some helpful links for getting started:
\itemize{
\item The R package documentation contains usage examples and method reference:
\item etc
}
}
\examples{
\donttest{
# The following script demonstrates...
}

}
\author{
\strong{Maintainer}: Yizhe Xu \email{yizhex@stanford.edu}

Authors:
\itemize{
  \item Nikolaos Ignatiadis
  \item Erik Sverdrup
  \item Scott Fleming
  \item Stefan Wager
  \item Nigam Shah
}

}
\keyword{internal}
