% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flasso.R
\name{Flasso}
\alias{Flasso}
\title{F-learner of lasso}
\usage{
Flasso(x, w, y, weight, pscore = rep(0.5, nrow(x)), nfolds = 10, alpha = 1)
}
\arguments{
\item{x}{matrix of covariates}

\item{w}{vector of treatment indicators (0 or 1)}

\item{y}{vector of response values}

\item{weight}{vector of subject level weights}

\item{pscore}{vector of propensity scores}

\item{nfolds}{number of cross-validation}

\item{alpha}{mixing tuning parameter for lasso}
}
\value{
An Flasso object
}
\description{
F-learner, implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
weight <- rep(1, length(Y))
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

Flasso_fit = Flasso(data$X, data$W, data$Y, weight = weight)
Flasso_cate = predict(Flasso_fit, data.test$X)
}
}
