% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgrf.R
\name{predict.rgrf}
\alias{predict.rgrf}
\title{predict for rgrf}
\usage{
\method{predict}{rgrf}(object, newx = NULL, tau_only = TRUE, ...)
}
\arguments{
\item{object}{a rgrf object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{tau_only}{if set to TRUE, onlly return prediction on tau. Otherwise, return a list including prediction on tau, propensity score, and baseline main effect.}

\item{...}{additional arguments (currently not used)}
}
\value{
A vector of predicted conditional average treatment effects
}
\description{
get estimated tau(x) using the trained rgrf model
}
\examples{
\donttest{
n <- 1000; p <- 25
times <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

rgrf_fit = rgrf(X, W, Y, D, times, p_hat = 0.5)
rgrf_cate = predict(rgrf_fit, X)
}

}
