% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoxph.R
\name{scoxph}
\alias{scoxph}
\title{S-learner of Cox PH}
\usage{
scoxph(x, w, y, D, times)
}
\arguments{
\item{x}{The baseline covariates}

\item{w}{The treatment variable (0 or 1)}

\item{y}{The follow-up time}

\item{D}{The event indicator}

\item{times}{The prediction time of interest}
}
\value{
a scoxph object
}
\description{
S-learner, implemented via Cox proportional hazard models
}
\examples{
\dontrun{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

scoxph_fit = scoxph(x, w, y, D, times)
scoxph_cate = predict(scoxph_fit, x, times)
}
}
