% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_sl_coxph.R
\name{surv_sl_coxph}
\alias{surv_sl_coxph}
\title{S-learner of Cox PH}
\usage{
surv_sl_coxph(X, W, Y, D, times)
}
\arguments{
\item{X}{The baseline covariates}

\item{W}{The treatment variable (0 or 1)}

\item{Y}{The follow-up time}

\item{D}{The event indicator}

\item{times}{The prediction time of interest}
}
\value{
a surv_sl_coxph object
}
\description{
S-learner, implemented via Cox proportional hazard models
}
\examples{
\donttest{
n = 1000; p = 25
times = 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)

surv_sl_coxph_fit = surv_sl_coxph(X, W, Y, D, times)
cate = predict(surv_sl_coxph_fit)
}
}
