% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_sl_grf.R
\name{surv_sl_grf}
\alias{surv_sl_grf}
\title{S-learner with random survival forest}
\usage{
surv_sl_grf(X, Y, W, D, t0, new.args.grf.nuisance = list())
}
\arguments{
\item{X}{The baseline covariates}

\item{Y}{The follow-up time}

\item{W}{The treatment variable (0 or 1)}

\item{D}{The event indicator}

\item{t0}{The prediction time of interest}

\item{new.args.grf.nuisance}{Input arguments for a grf model that estimates nuisance parameters}
}
\value{
A surv_sl_grf object
}
\description{
Estimating conditional average treatment effects (CATEs) for
survival outcomes using S-learner with random survival forest predictive models
(implemented via the grf package).
The CATE is defined as tau(X) = p(Y(1) > t0 | X = x) - p(Y(0) > t0 | X = x),
where Y(1) and Y(0) are counterfactual survival times under the treated and controlled arms, respectively.
}
\examples{
\donttest{
n <- 1000; p <- 25
t0 <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[ ,1, drop = FALSE] + (-0.5 - 1 * X[ ,2, drop = FALSE]) * W)) ^ 2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC / (4 ^ 2)) ^ (1 / 2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.sl.grf.fit <- surv_sl_grf(X, Y, W, D, t0)
cate <- predict(surv.sl.grf.fit)
cate.test <- predict(surv.sl.grf.fit, X.test)
}
}
