% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_tl_coxph.R
\name{surv_tl_coxph}
\alias{surv_tl_coxph}
\title{T-learner of Cox PH}
\usage{
surv_tl_coxph(data, data.test, times)
}
\arguments{
\item{data}{The training data set}

\item{data.test}{The testing data set}

\item{times}{The prediction time of interest}
}
\value{
A vector of estimated conditional average treatment effects
}
\description{
T-learner, implemented via Cox proportional hazard models
}
\examples{
\donttest{
n <- 1000; p <- 25
times <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[,1] + (-0.5 - 1 * X[,2]) * W))^2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC/(4^2))^(1/2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
data <- list(X = X, W = W, Y = Y, D = D)
data.test <- list(X = X, W = W, Y = Y, D = D)

cate = surv_tl_coxph(data, data.test, times)
}
}
