% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_tl_lasso.R
\name{surv_tl_lasso}
\alias{surv_tl_lasso}
\title{T-learner of lasso}
\usage{
surv_tl_lasso(X, Y, W, D, t0, k.folds = 10)
}
\arguments{
\item{X}{The baseline covariates}

\item{Y}{The follow-up time}

\item{W}{The treatment variable (0 or 1)}

\item{D}{The event indicator}

\item{t0}{The prediction time of interest}

\item{k.folds}{The number of folds for estimating nuisance parameters via cross-fitting}
}
\value{
A surv_tl_lasso object
}
\description{
T-learner, implemented via glmnet (lasso)
}
\examples{
\donttest{
n <- 1000; p <- 25
t0 <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[ ,1] + (-0.5 - 1 * X[ ,2]) * W)) ^ 2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC / (4 ^ 2)) ^ (1 / 2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.tl.lasso.fit <- surv_tl_lasso(X, Y, W, D, t0)
cate <- predict(surv.tl.lasso.fit)
cate.test <- predict(surv.tl.lasso.fit, X.test)
}
}
