% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_xl_lasso.R
\name{surv_xl_lasso}
\alias{surv_xl_lasso}
\title{X-learner of lasso}
\usage{
surv_xl_lasso(
  X,
  Y,
  W,
  D,
  t0,
  alpha = 0.05,
  W.hat = NULL,
  cen.fit = "Kaplan-Meier",
  k.folds = 10
)
}
\arguments{
\item{X}{The baseline covariates}

\item{Y}{The follow-up time}

\item{W}{The treatment variable (0 or 1)}

\item{D}{The event indicator}

\item{t0}{The prediction time of interest}

\item{alpha}{Imbalance tuning parameter for a split (see grf documentation)}

\item{W.hat}{The propensity score}

\item{cen.fit}{The choice of model fitting for censoring}

\item{k.folds}{The number of folds for estimating nuisance parameters via cross-fitting}
}
\value{
A surv_xl_lasso object
}
\description{
X-learner, implemented via glmnet (lasso) with "coxph" distribution
}
\examples{
\donttest{
n <- 1000; p <- 25
t0 <- 0.2
Y.max <- 2
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
numeratorT <- -log(runif(n))
T <- (numeratorT / exp(1 * X[ ,1] + (-0.5 - 1 * X[ ,2]) * W)) ^ 2
failure.time <- pmin(T, Y.max)
numeratorC <- -log(runif(n))
censor.time <- (numeratorC / (4 ^ 2)) ^ (1 / 2)
Y <- pmin(failure.time, censor.time)
D <- as.integer(failure.time <= censor.time)
n.test <- 500
X.test <- matrix(rnorm(n.test * p), n.test, p)

surv.xl.lasso.fit <- surv_xl_lasso(X, Y, W, D, t0, W.hat = 0.5)
cate <- predict(surv.xl.lasso.fit)
cate.test <- predict(surv.xl.lasso.fit, X.test)
}
}
